requirements.txt
asyncpg==0.30.0
fastapi==0.115.8
httpx==0.28.1
minio==7.2.15
openai==1.65.2
pdfplumber==0.11.5
pgvector==0.3.6
Pillow==11.1.0
pydantic-settings==2.7.1
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.25.3
SQLAlchemy==2.0.38
tiktoken==0.9.0
uvicorn==0.34.0
